+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   `@lumx/react`:
+    -   `Slider`: support dragging the handle on touch devices.
+
 ## [4.20.0][] - 2026-07-24
 
 ### Added